+++ src/lib/networkHandler.js
@@ -44,6 +44,7 @@ class NetworkHandler {
         this.bigIp = bigIp;
         this.eventEmitter = eventEmitter;
         this.state = state;
+        this.needsMcpdRestart = false;
     }
 
     /**
@@ -85,6 +86,14 @@ class NetworkHandler {
                 return handleFirewallPolicy.call(this);
             })
             .then(() => {
+                logger.fine('Checking Net Address Lists');
+                return handleNetAddressList.call(this);
+            })
+            .then(() => {
+                logger.fine('Checking Net Port Lists');
+                return handleNetPortList.call(this);
+            })
+            .then(() => {
                 logger.fine('Checking ManagementIpFirewall.');
                 return handleManagementIpFirewall.call(this);
             })
@@ -122,6 +131,40 @@ class NetworkHandler {
             })
             .then(() => {
                 logger.info('Done processing network declartion.');
+                return Promise.resolve();
+            })
+            .then(() => {
+                if (this.needsMcpdRestart) {
+                    logger.info('Saving config');
+                    return this.bigIp.save();
+                }
+                return Promise.resolve();
+            })
+            .then(() => {
+                if (this.needsMcpdRestart) {
+                    logger.info('Restarting mcpd');
+                    const servicesToWaitFor = [
+                        'cbrd',
+                        'alertd',
+                        'tamd',
+                        'lind',
+                        'lacpd',
+                        'merged',
+                        'devmgmtd',
+                        'pccd',
+                        'logstatd',
+                        'statsd',
+                        'icr_eventd',
+                        'zxfrd',
+                        'wccpd',
+                        'tmm',
+                        'vxland',
+                        'dynconfd',
+                        'named',
+                        'bigd'
+                    ];
+                    return doUtil.restartService(this.bigIp, 'mcpd', servicesToWaitFor);
+                }
                 return Promise.resolve();
             })
             .catch((err) => {
@@ -288,6 +331,56 @@ function handleFirewallPortList() {
     return Promise.all(promises)
         .catch((err) => {
             logger.severe(`Error creating Firewall Port List: ${err.message}`);
+            throw err;
+        });
+}
+
+function handleNetAddressList() {
+    const promises = [];
+
+    doUtil.forEach(this.declaration, 'NetAddressList', (tenant, netAddressList) => {
+        if (netAddressList && netAddressList.name) {
+            const body = {
+                name: netAddressList.name,
+                description: netAddressList.description
+            };
+
+            if (netAddressList.addresses) {
+                body.addresses = netAddressList.addresses || [];
+            }
+
+            promises.push(this.bigIp.createOrModify(PATHS.NetAddressList, body, null, cloudUtil.MEDIUM_RETRY));
+        }
+    });
+
+    return Promise.all(promises)
+        .catch((err) => {
+            logger.severe(`Error creating Net Address List: ${err.message}`);
+            throw err;
+        });
+}
+
+function handleNetPortList() {
+    const promises = [];
+
+    doUtil.forEach(this.declaration, 'NetPortList', (tenant, netPortList) => {
+        if (netPortList && netPortList.name) {
+            const body = {
+                name: netPortList.name,
+                description: netPortList.description
+            };
+
+            if (netPortList.ports) {
+                body.ports = netPortList.ports.map((port) => port.toString());
+            }
+
+            promises.push(this.bigIp.createOrModify(PATHS.NetPortList, body, null, cloudUtil.MEDIUM_RETRY));
+        }
+    });
+
+    return Promise.all(promises)
+        .catch((err) => {
+            logger.severe(`Error creating Net Port List: ${err.message}`);
             throw err;
         });
 }
@@ -618,6 +711,7 @@ function handleRouteDomain() {
             if (this.state.currentConfig.Common.RouteDomain
                 && this.state.currentConfig.Common.RouteDomain[routeDomain.name]) {
                 method = 'modify';
+                this.needsMcpdRestart = true;
             }
             commands.push({
                 method,