+++ test/unit/schema/networkSchemaTests.js
@@ -352,6 +352,17 @@ describe('network.schema.json', () => {
                 };
                 assert.ok(validate(data), getErrorString(validate));
             });
+
+            it('should validate network data with allow service:port and default', () => {
+                const data = {
+                    class: 'SelfIp',
+                    address: '1.2.3.4/32',
+                    vlan: 'myVlan',
+                    allowService: ['foo:1234', 'default'],
+                    trafficGroup: 'traffic-group-1'
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
         });
 
         describe('invalid', () => {
@@ -450,6 +461,17 @@ describe('network.schema.json', () => {
                     };
                     assert.strictEqual(validate(data), false, 'allow service bar should not be valid port');
                     assert(getErrorString().includes('should match pattern'));
+                });
+
+                it('should invalidate misspelled default in port array', () => {
+                    const data = {
+                        class: 'SelfIp',
+                        address: '1.2.3.4',
+                        vlan: 'myVlan',
+                        allowService: ['foo:1234', 'defalt']
+                    };
+                    assert.strictEqual(validate(data), false, 'default should be spelled correctly');
+                    assert(getErrorString().includes('should match pattern \\"(\\\\w+:\\\\d+|default)\\"'));
                 });
             });
         });
@@ -2412,6 +2434,77 @@ describe('network.schema.json', () => {
             it('should invalidate if no ports are specified', () => {
                 const data = {
                     class: 'FirewallPortList'
+                };
+                assert.strictEqual(validate(data), false);
+                assert(getErrorString().includes('should have required property'));
+            });
+        });
+    });
+
+    describe('NetAddressList', () => {
+        describe('valid', () => {
+            it('should validate minimal net address list properties', () => {
+                const data = {
+                    class: 'NetAddressList',
+                    addresses: ['192.168.0.1']
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate all net address list properties', () => {
+                const data = {
+                    class: 'NetAddressList',
+                    remark: 'myRemark',
+                    addresses: ['192.168.0.1']
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+        });
+
+        describe('invalid', () => {
+            it('should invalidate if no addresses are specified', () => {
+                const data = {
+                    class: 'NetAddressList'
+                };
+                assert.strictEqual(validate(data), false);
+                assert(getErrorString().includes('should have required property'));
+            });
+        });
+    });
+
+    describe('NetPortList', () => {
+        describe('valid', () => {
+            it('should validate minimal net port list properties', () => {
+                const data = {
+                    class: 'NetPortList',
+                    ports: [8080]
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate all net port list properties', () => {
+                const data = {
+                    class: 'NetPortList',
+                    remark: 'myRemark',
+                    ports: [8080]
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate net port list with string port', () => {
+                const data = {
+                    class: 'NetPortList',
+                    remark: 'myRemark',
+                    ports: ['8080']
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+        });
+
+        describe('invalid', () => {
+            it('should invalidate if no ports are specified', () => {
+                const data = {
+                    class: 'NetPortList'
                 };
                 assert.strictEqual(validate(data), false);
                 assert(getErrorString().includes('should have required property'));